Switch over to new Color struct as default.
Former-commit-id: 36ee757ac11c4a801317a501ff723424c2cb37ce
Former-commit-id: d2c80a5c23eb7dbc38f4384ea5c9497d48582fdf
Former-commit-id: 02e23e7442f61e89f3e2d903be05f11ecce509c7

+++ src/ImageProcessor/ParallelImageProcessor.cs
@@ -53,7 +53,7 @@ namespace ImageProcessor
         /// <inheritdoc/>
         public void Apply(ImageBase target, ImageBase source, int width, int height, Rectangle targetRectangle = default(Rectangle), Rectangle sourceRectangle = default(Rectangle))
         {
-            byte[] pixels = new byte[width * height * 4];
+            float[] pixels = new float[width * height * 4];
             target.SetPixels(width, height, pixels);
 
             if (targetRectangle == Rectangle.Empty)
